Refactor implementation part and add an HttpInterceptor
