+++ daikon-spring/daikon-spring-audit-logs/src/main/java/org/talend/daikon/spring/audit/logs/exception/AuditLogException.java
@@ -1,0 +1,12 @@
+package org.talend.daikon.spring.audit.logs.exception;
+
+public class AuditLogException extends Exception {
+
+    public AuditLogException(String message) {
+        super(message);
+    }
+
+    public AuditLogException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
