--- daikon-spring/daikon-spring-audit-logs/src/main/java/org/talend/daikon/spring/audit/logs/service/AuditLogGenerationFilter.java
@@ -1,11 +1,0 @@
-package org.talend.daikon.spring.audit.logs.service;
-
-import org.talend.daikon.spring.audit.logs.api.AuditUserProvider;
-import org.talend.logging.audit.Context;
-
-public interface AuditLogGenerationFilter {
-
-    void sendAuditLog(Context context);
-
-    AuditUserProvider getAuditUserProvider();
-}
