--- daikon-spring/daikon-spring-audit-logs/src/main/java/org/talend/daikon/spring/audit/logs/service/AuditLogGenerationFilterInterface.java
@@ -1,8 +1,0 @@
-package org.talend.daikon.spring.audit.logs.service;
-
-import org.talend.logging.audit.Context;
-
-public interface AuditLogGenerationFilterInterface {
-
-    void sendAuditLog(Context context);
-}
